+++ udata_front/theme/gouvfr/assets/js/components/Report/ReportModalButton.vue
@@ -17,7 +17,7 @@
     />
 </template>
 <script setup lang="ts">
-import { getRandomId } from '@etalab/data.gouv.fr-components';
+import { getRandomId } from '@datagouv/components';
 import { ref } from 'vue';
 import { useI18n } from 'vue-i18n';
 import ReportModal, { type ReportModalForm } from './ReportModal.vue';